# swap babel and ts-node template compilation for swc