--- a/.changeset/smart-singers-draw.md
+++ b/.changeset/smart-singers-draw.md
@@ -1,0 +1,7 @@
+---
+"mailing": patch
+"mailing-core": patch
+"web": patch
+---
+
+swap babel and ts-node template compilation for swc