--- a/packages/cli/package.json
+++ b/packages/cli/package.json
@@ -26,6 +26,8 @@
   "description": "Fun email development environment",
   "devDependencies": {
     "@next/eslint-plugin-next": "^12.2.0",
+    "@swc/core": "^1.2.229",
+    "@swc/register": "^0.1.10",
     "@types/fs-extra": "^9.0.13",
     "@types/html-minifier-terser": "^6.1.0",
     "@types/mjml": "^4.7.0",
@@ -42,10 +44,6 @@
     "react-dom": "^17.0.0 || ^18.0.0"
   },
   "dependencies": {
-    "@babel/core": "^7.18.9",
-    "@babel/preset-env": "^7.18.6",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/register": "^7.18.9",
     "chalk": "^4.1.2",
     "dotenv": "^16.0.1",
     "fs-extra": "^10.1.0",
@@ -59,7 +57,6 @@
     "react-hotkeys-hook": "^3.4.6",
     "react-tiny-popover": "^7.1.0",
     "tree-node-cli": "^1.5.2",
-    "ts-node": "^10.7.0",
     "typescript": "^4.6.3",
     "uglify-js": "^3.15.4",
     "yargs": "^17.5.1"