--- a/packages/cli/src/.swcrc
+++ b/packages/cli/src/.swcrc
@@ -1,0 +1,18 @@
+{
+  "jsc": {
+    "target": "es2015",
+    "parser": {
+      "syntax": "typescript",
+      "tsx": true,
+      "decorators": true
+    },
+    "transform": {
+      "react": {
+        "runtime": "automatic"
+      }
+    }
+  },
+  "module": {
+    "type": "commonjs"
+  }
+}