--- a/packages/cli/src/commands/util/registerRequireHooks.ts
+++ b/packages/cli/src/commands/util/registerRequireHooks.ts
@@ -1,25 +1,4 @@
 export default function registerRequireHooks() {
   if (process.env.MM_DEV) return;
-
-  require("ts-node").register({
-    compilerOptions: {
-      module: "commonjs",
-      jsx: "react-jsx",
-      moduleResolution: "node",
-      skipLibCheck: true,
-    },
-  });
-
-  require("@babel/register")({
-    presets: [
-      [
-        "@babel/react",
-        {
-          runtime: "automatic",
-        },
-      ],
-      "@babel/preset-env",
-    ],
-    compact: false,
-  });
+  require("@swc/register");
 }