--- a/packages/cli/src/commands/util/registerRequireHooks.ts
+++ b/packages/cli/src/commands/util/registerRequireHooks.ts
@@ -1,4 +1,20 @@
 export default function registerRequireHooks() {
   if (process.env.MM_DEV) return;
-  require("@swc/register");
+  require("@swc/register")({
+    jsc: {
+      parser: {
+        syntax: "typescript",
+        tsx: true,
+        decorators: true,
+      },
+      transform: {
+        react: {
+          runtime: "automatic",
+        },
+      },
+    },
+    module: {
+      type: "commonjs",
+    },
+  });
 }