--- a/packages/cli/src/dev.js
+++ b/packages/cli/src/dev.js
@@ -3,5 +3,21 @@
 // This script can be used for quick cli development without compilation steps.
 
 process.env.MM_DEV = 1;
-require("@swc/register");
+require("@swc/register")({
+  jsc: {
+    parser: {
+      syntax: "typescript",
+      tsx: true,
+      decorators: true,
+    },
+    transform: {
+      react: {
+        runtime: "automatic",
+      },
+    },
+  },
+  module: {
+    type: "commonjs",
+  },
+});
 require("./index.ts");