fixes #102 

This makes the initial index request much faster (seems like 2-3x). Here's the export-previews command that had the same slow template compilation. I generated these benchmarks with hyperfine.


### TS

Before...
```
Benchmark 1: npx mailing export-previews
  Time (mean ± σ):      4.804 s ±  0.057 s    [User: 7.017 s, System: 0.649 s]
  Range (min … max):    4.720 s …  4.878 s    10 runs
```

After...
```
Benchmark 1: npx mailing export-previews
  Time (mean ± σ):      1.779 s ±  0.044 s    [User: 1.503 s, System: 0.424 s]
  Range (min … max):    1.717 s …  1.855 s    10 runs
```


### JS

Before...
```
Benchmark 1: npx mailing export-previews
  Time (mean ± σ):      2.497 s ±  0.118 s    [User: 2.290 s, System: 0.524 s]
  Range (min … max):    2.337 s …  2.712 s    10 runs
```

After...
```
Benchmark 1: npx mailing export-previews
  Time (mean ± σ):      1.704 s ±  0.042 s    [User: 1.443 s, System: 0.406 s]
  Range (min … max):    1.664 s …  1.814 s    10 runs
```